+++ src/DependencyInjection/Configuration.php
@@ -29,12 +29,7 @@ final class Configuration implements ConfigurationInterface
         $rootNode
             ->children()
                 ->scalarNode('dsn')
-                    ->beforeNormalization()
-                        ->ifTrue(static function ($value): bool {
-                            return empty($value) || (\is_string($value) && '' === trim($value));
-                        })
-                        ->thenUnset()
-                    ->end()
+                    ->info('If this value is not provided, the SDK will try to read it from the SENTRY_DSN environment variable. If that variable also does not exist, the SDK will just not send any events.')
                 ->end()
                 ->booleanNode('register_error_listener')->defaultTrue()->end()
                 ->arrayNode('options')